requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.3
geographiclib==1.52
geojson==2.5.0
geomet==0.3.0
geopy==2.2.0
idna==3.2
iniconfig==1.1.1
numpy==1.21.2
packaging==21.0
pluggy==1.0.0
py==1.10.0
pymongo==4.0.1
pyparsing==2.4.7
pyproj==3.2.1
pytest==6.2.5
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.7
